+++ src/storyteller/handler.rs
@@ -7,74 +7,16 @@ use std::sync::{Arc, Mutex, RwLock};
 use std::{io, marker};
 use storyteller::{EventHandler, Reporter};
 
-pub trait SendWriter: io::Write + Send + 'static {}
+mod discard_output_handler;
+mod human_progress_handler;
+mod json_handler;
 
-pub struct JsonHandler<W: SendWriter> {
-    writer: Arc<Mutex<W>>,
-}
+#[cfg(test)]
+mod testing_handler;
 
-impl<W: SendWriter> JsonHandler<W> {
-    const LOCK_FAILURE_MSG: &'static str =
-        "{ \"panic\": true, \"cause\": \"Unable to lock writer for JsonHandle\", \"experimental\": true }";
-    const SERIALIZE_FAILURE_MSG: &'static str =
-        "{ \"panic\": true, \"cause\": \"Unable to serialize event for JsonHandle\", \"experimental\": true }";
-    const WRITE_FAILURE_MSG: &'static str =
-        "{ \"panic\": true, \"cause\": \"Unable to write serialized event for JsonHandle\", \"experimental\": true }";
-}
+pub use discard_output_handler::DiscardOutputHandler;
+pub use human_progress_handler::HumanProgressHandler;
+pub use json_handler::JsonHandler;
 
-impl SendWriter for Stderr {}
-
-impl JsonHandler<Stderr> {
-    pub fn stderr() -> Self {
-        Self {
-            writer: Arc::new(Mutex::new(io::stderr())),
-        }
-    }
-}
-
-impl<W: SendWriter> EventHandler for JsonHandler<W> {
-    type Event = super::Event;
-
-    fn handle(&self, event: Self::Event) {
-        let mut w = self.writer.lock().expect(Self::LOCK_FAILURE_MSG);
-        let serialized_event = serde_json::to_string(&event).expect(Self::SERIALIZE_FAILURE_MSG);
-
-        writeln!(&mut w, "{}", &serialized_event).expect(Self::WRITE_FAILURE_MSG);
-    }
-}
-
-pub struct HumanProgressHandler {
-    bar: indicatif::ProgressBar,
-}
-
-impl HumanProgressHandler {
-    pub fn new() -> Self {
-        Self {
-            bar: indicatif::ProgressBar::new(0),
-        }
-    }
-}
-
-impl EventHandler for HumanProgressHandler {
-    type Event = super::Event;
-
-    fn handle(&self, event: Self::Event) {
-        match event {
-            Event::Todo(msg) => self.bar.println(msg),
-            Event::Progress(progress) => {}
-            Event::Action(action) => {
-                self.bar.println(Into::<String>::into(action));
-            }
-        }
-    }
-}
-
-// ---------------
-
-pub struct DiscardOutputHandler;
-
-impl EventHandler for DiscardOutputHandler {
-    type Event = super::Event;
-
-    fn handle(&self, _event: Self::Event) {}
-}
+#[cfg(test)]
+pub use testing_handler::TestingHandler;
