+++ src/storyteller/handler/discard_output_handler.rs
@@ -1,0 +1,9 @@
+use storyteller::EventHandler;
+
+pub struct DiscardOutputHandler;
+
+impl EventHandler for DiscardOutputHandler {
+    type Event = super::Event;
+
+    fn handle(&self, _event: Self::Event) {}
+}
